+++ frontend/components/ExamContent.tsx
@@ -2,8 +2,8 @@ import { FC, useCallback, useEffect, useRef, useState } from "react";
 import Router from "next/router";
 import { ChevronDownIcon } from "@heroicons/react/solid";
 import CodeArea from "./CodeArea";
-import QuestionArea2 from "./QuestionArea2";
-import Timer2 from "./Timer2";
+import QuestionArea2 from "./QuestionArea";
+import Timer2 from "./Timer";
 import InstructionModal from "./InstructionModal";
 import AlertModal from "./AlertModal";
 import useFullScreen from "../hooks/useFullScreen";
