--- frontend/declaration.d.ts
@@ -1,5 +1,0 @@
-declare namespace JSX {
-    interface IntrinsicElements {
-      "lottie-player": any;
-    }
-  }
