+++ frontend/package.json
@@ -11,16 +11,10 @@
   "dependencies": {
     "@azure/msal-browser": "^2.22.1",
     "@azure/msal-react": "^1.3.1",
-    "@emotion/cache": "^11.7.1",
-    "@emotion/react": "^11.8.2",
-    "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.8.1",
     "@headlessui/react": "^1.5.0",
     "@heroicons/react": "^1.0.6",
     "@lottiefiles/lottie-player": "^1.5.6",
     "@monaco-editor/react": "^4.3.1",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
     "@twilio/conversations": "^2.1.0",
     "axios": "^0.26.1",
     "dayjs": "^1.11.0",
@@ -29,9 +23,9 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-markdown": "^8.0.1",
+    "react-toastify": "^8.2.0",
     "recharts": "^2.1.9",
-    "twilio": "^3.76.1",
-    "react-toastify": "^8.2.0"
+    "twilio": "^3.76.1"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.0",
