+++ frontend/pages/_app.tsx
@@ -1,48 +1,23 @@
 import * as React from "react";
-import Head from "next/head";
 import { AppProps } from "next/app";
-import { CacheProvider, EmotionCache } from "@emotion/react";
-import createEmotionCache from "../src/createEmotionCache";
 import { PublicClientApplication } from "@azure/msal-browser";
 import { MsalProvider } from "@azure/msal-react";
 import { msalConfig } from "../src/authConfig";
-import { StyledEngineProvider } from "@mui/material/styles";
 import "../styles/globals.css";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.min.css";
 import UserProvider from "../context/UserProvider";
-// Client-side cache, shared for the whole session of the user in the browser.
-const clientSideEmotionCache = createEmotionCache();
 
-interface MyAppProps extends AppProps {
-  emotionCache?: EmotionCache;
-}
-export default function MyApp(props: MyAppProps) {
+function MyApp({ Component, pageProps }: AppProps) {
   const msalInstance = new PublicClientApplication(msalConfig);
-  const { Component, emotionCache = clientSideEmotionCache, pageProps } = props;
-
-  React.useEffect(() => {
-    // Remove the server-side injected CSS.
-    const jssStyles = document.querySelector("#jss-server-side");
-    if (jssStyles) {
-      jssStyles.parentElement?.removeChild(jssStyles);
-    }
-  }, []);
-
   return (
-    <CacheProvider value={emotionCache}>
-      <Head>
-        <meta name="viewport" content="initial-scale=1, width=device-width" />
-      </Head>
-      {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-      <StyledEngineProvider injectFirst>
-        <MsalProvider instance={msalInstance}>
-          <UserProvider>
-            <Component {...pageProps} />
-            <ToastContainer />
-          </UserProvider>
-        </MsalProvider>
-      </StyledEngineProvider>
-    </CacheProvider>
+    <MsalProvider instance={msalInstance}>
+      <UserProvider>
+        <Component {...pageProps} />
+        <ToastContainer />
+      </UserProvider>
+    </MsalProvider>
   );
 }
+
+export default MyApp;
